Improve slurm-spank-plugins build system for build outside LLNL.
Rename project to slurm-spank-plugins and create
slurm-spank-plugins.spec specfile with new cpuset and llnl
subpackages. To build these subpackages, use '--with cpuset' and/or
'--with llnl_plugins' on the rpmbuild command line.

These changes are just an incremental improvement over previous
build system, but should suffice until the project is converted to
autoconf & friends.