preserve-env: Fix crash in slurmstepd when env vars longer than 64B.
A very small buffer was accidentally used in preserve-env.so when
dealing with environment variables from the job's environment. Make
all env var buffers dynamic (except for the name of save_SLURM_*
variables) to avoid this problem in the future.

The fix also includes an undocumented --preserve-slurm-env=verbose
debug mode, to make it easier to debug future problems with the plugin.