use-env: fix potential segv if spank exit hook called before init
In some cases it is possible that the spank_exit hook will be
called without the spank_init function having been called. This
can happen for instance if a previous required plugin fails to
initialize.

The use-env plugin had the potential to cause a segv during
cleanup of the log facility due to a bad free() of log_ctx.prefix.
This patch fixes that case.