Move slurmstepd process to user cpuset after all tasks exit
In some cases, the slurmstepd process may hang around for many
seconds after SLURM has marked a job and/or  job step fully complete.
This causes problems for the cpuset code becuase the slurmstepd
process is holding the job cpuset, and the code thinks that the
associated CPUs are still in use. This causes failures to run
new jobs as it appears there are not enough CPUs.

To prevent this failure mode, add a spank_task_exit hook in the
cpuset plugin which moves the slurmstepd process back to the user
cpuset after all tasks have exited. This will prevent the rest
of the code from accounting these CPUs as in-use.