Remove unecessary calls to dynamically loaded slurm symbols
The SLURM API (libslurm.so) must be dlopened to ensure that RTLD_GLOBAL
is passed so that further plugins loaded by slurm calls have access to
SLURM symbols. However, it is safer to link with libslurm and use the
slurm api calls directly instead of the dlsym provided symbols, so
switch to using that method.