lua: Avoid lua stack overflow
Ensure that all functions clean up lua stack on return
to avoid the possibility of unecessarily overflowing the
stack.

The specific case fixed here was a nil left on the stack
in lua_spank_call() when the function to be called did not
exist in the current lua script. If lua_spank_call was
called repeatedly, the lua stack would overflow and cause
a segv.