auto-affinity: Check for SLURM cpus per task option
In auto-affinity the SLURM cpus_per_task setting is currently
ignored, because there is no way to tell if the user explictly
set the value when cpus_per_task == 1.  This patch assumes that,
if cpus_per_task as returned by S_STEP_CPUS_PER_TASK is 2 or greater
the user explicitly requested this value from SLURM, and the
auto-affinity module should behave as if called as

 --auto-affinity=cpt=N

Thus

 srun -c 2 -n 5 args...

would bind each of 5 tasks to two CPUs on a 12 CPU system, instead
of giving up like it does now.