auto-affinity: Enable CPU affinity anytime ncpus/ntasks greater than 1
Currenlty, the auto-affinity plugin does not adjust the CPU affinity
mask if the number of available CPUs on the node is not an even multiple
of the number of tasks. This patch relieves that restriction a bit
by assuming that the user wanted to have the next multiple of CPUs
per task assigned to each task, iff nCPUs/ntasks > 1. For example,
if 5 tasks are launched are launched on a 12 CPU node, the new code
will assign 2 CPUs per task by default, but if 7 tasks were launched,
the code would not adjust affinity masks.

The old behavior can be restored by adding the option "multiples_only"
to the plugstack.conf line for auto-affinity.so.

This patch also moves these checks for divisibility into spank_user_init()
so that they are performed only once per job step and not once per task.