auto-affinity: Fix tests for enabling CPU affinity
Previous patch to enable CPU affinity any time ncpus/ntasks > 1
broke the auto-enabling of CPU affinity for the case where the
'multiples_only' option is not set.

This patch cleans up the code in slurm_spank_user_init() to make it
more clear when CPU affinity should be automatically enabled.