use-env: Fix segfault in remote option processing
During spank_init() on the remote side, the use-env plugin initializes
the environment operations that use the spank handle for later use.
The spank_t object is passed in as the context arg and is later used
by all x*env() functions.

This apparently used to work when the x*env() calls were used during
the spank option callbacks. However, in a recent version of SLURM
the spank_t handle becomes invalid between spank_init() and the later
option callback (use_env_opt_process()). This causes the call to
xgetenv("HOME") to generate a segfault in slurmstepd.

Therefore, we need to initialize all environment vars we'll need later
in spank_init(). For now this means using a global reference for HOME,
which we use later in env_override_file_search(), which is called
from use_env_opt_process().