use-env: don't fail on missing HOME env var
The use-env plugin was unnecessarily failing when the HOME
environment variable was not defined. Specifically, when
submitting jobs to the Moab scheduler, it would run sbatch(1)
with an empty environment, and thus HOME was missing. Given
that use-env is likely not useful in this scenario, instead
of bailing, just disable the plugin.