mpibind:  Modify the distribution of threads to proc units
This change only applies to the case when the number of software
threads is less than the number of cores allotted to a task.
Formally, we distributed one thread to a hardware thread when SMT was
enabled and two threads could be assigned to the same core.  Now we
distribute one thread to a core.  Addresses a problem identified as
part of the discussion in TOSS-3338.